create separate types directory
