+++ src/components/UserInfo/UserInfo.tsx
@@ -1,5 +1,5 @@
 import { FC } from 'react';
-import { User } from '../../react-app-env';
+import { User } from '../../types/User';
 
 type Props = {
   user: User;
