+++ src/react-app-env.d.ts
@@ -1,18 +1,2 @@
 /// <reference types="react-scripts" />
-export interface Todo {
-  id: number,
-  title: string,
-  completed: boolean,
-  userId: number,
-}
 
-export interface User {
-  id: number,
-  name: string,
-  username: string,
-  email: string,
-}
-
-export interface ExtendedTodo extends Todo {
-  user: User | null;
-}
