+++ src/types/ExtendedTodo.ts
@@ -1,0 +1,6 @@
+import {User} from "./User";
+import {Todo} from "./Todo";
+
+export interface ExtendedTodo extends Todo {
+  user: User | null;
+}
